Apply some corrections from code review
