+++ .gitignore
@@ -12,5 +12,5 @@ _trial_temp.lock
 # Runtime files
 /sydent.conf
 /sydent.db
+/sydent.pid
 /matrix_is_test/sydent.stderr
-sydent.pid
