--- sydent/config/_base.py
@@ -1,2 +1,0 @@
-class BaseConfig:
-    pass
