+++ sydent/config/crypto.py
@@ -1,16 +1,29 @@
+# Copyright 2021 The Matrix.org Foundation C.I.C.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
 import logging
 from configparser import ConfigParser
 
 import nacl
 import signedjson.key
 
-from sydent.config._base import BaseConfig
-
 logger = logging.getLogger(__name__)
 
 
-class CryptoConfig(BaseConfig):
-    def parse_config(self, cfg: ConfigParser):
+class CryptoConfig:
+    def parse_config(self, cfg: ConfigParser) -> None:
         """
         Parse the crypto section of the config
