--- sydent/sign/ed25519.py
@@ -1,26 +1,0 @@
-# Copyright 2014 OpenMarket Ltd
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from sydent.sydent import Sydent
-
-
-# TODO just remove this class
-class SydentEd25519:
-    def __init__(self, syd: "Sydent"):
-        self.sydent = syd
-
-        self.signing_key = self.sydent.config.crypto.signing_key
