+++ sydent/sydent.py
@@ -22,7 +22,7 @@ import os
 import twisted.internet.reactor
 from twisted.internet import address, task
 
-from sydent.config.server import SydentConfig
+from sydent.config import SydentConfig
 from sydent.db.hashing_metadata import HashingMetadataStore
 from sydent.db.sqlitedb import SqliteDatabase
 from sydent.db.valsession import ThreePidValSessionStore
@@ -85,6 +85,21 @@ class Sydent:
         self.config = cfg
         self.reactor = reactor
         self.use_tls_for_federation = use_tls_for_federation
+
+        if self.config.general.prometheus_enabled:
+            import prometheus_client
+
+            prometheus_client.start_http_server(
+                port=self.config.general.prometheus_port,
+                addr=self.config.general.prometheus_addr,
+            )
+
+        if self.config.general.sentry_enabled:
+            import sentry_sdk
+
+            sentry_sdk.init(dsn=self.config.general.sentry_dsn)
+            with sentry_sdk.configure_scope() as scope:
+                scope.set_tag("sydent_server_name", self.config.general.server_name)
 
         self.db = SqliteDatabase(self).db
 
@@ -226,19 +241,15 @@ class Sydent:
 
     def get_branded_template(self, brand, template_name):
         """
-        Calculate a (maybe) branded template filename to use.
+        Calculate a branded template filename to use.
 
-        If the deprecated email.template setting is defined, always use it.
-        Otherwise, attempt to use the hinted brand from the request if the brand
+        Attempt to use the hinted brand from the request if the brand
         is valid. Otherwise, fallback to the default brand.
 
         :param brand: The hint of which brand to use.
         :type brand: str or None
         :param template_name: The name of the template file to load.
         :type template_name: str
-        :param deprecated_template_name: The deprecated setting to use, if provided.
-        :type deprecated_template_name: Tuple[str]
-
         :return: The template filename to use.
         :rtype: str
         """
